[#69] refact(chart) - BarChartView, PiChartView 코드 분리
